+++ lib/elastic_apm/error/exception.rb
@@ -6,16 +6,21 @@ module ElasticAPM
     class Exception
       MOD_SPLIT = '::'
 
-      def initialize(exception, **attrs)
-        @message =
-          "#{exception.class}: #{exception.message}"
-        @type = exception.class.to_s
-        @module = format_module exception
-        @cause = exception.cause && Exception.new(exception.cause)
+      def initialize(attrs = nil)
+        return unless attrs
 
         attrs.each do |key, val|
           send(:"#{key}=", val)
         end
+      end
+
+      def self.from_exception(exception, **attrs)
+        new({
+          message: exception.message.to_s,
+          type: exception.class.to_s,
+          module: format_module(exception),
+          cause: exception.cause && Exception.from_exception(exception.cause)
+        }.merge(attrs))
       end
 
       attr_accessor(
@@ -29,10 +34,12 @@ module ElasticAPM
         :cause
       )
 
-      private
+      class << self
+        private
 
-      def format_module(exception)
-        exception.class.to_s.split(MOD_SPLIT)[0...-1].join(MOD_SPLIT)
+        def format_module(exception)
+          exception.class.to_s.split(MOD_SPLIT)[0...-1].join(MOD_SPLIT)
+        end
       end
     end
   end