+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Breaking Changes
+
+#### Removed
+
+- Removed all deprecated methods and config options
+
+#### Changed
+
+The following changes are breaking, as they may change the way data is grouped in Kibana.
+No changes are necessary to your app.
+
+- Errors' `message` no longer include their `type` ([#323](https://github.com/elastic/apm-agent-ruby/pull/323/files))
+- External request spans now have type `external.http.{library}` ([#514](https://github.com/elastic/apm-agent-ruby/pull/514))
+
 ## 2.12.0 (2019-10-01)
 
 ### Changed